changed seedPercentage datatype from int to double

+++ src/main/java/com/swiggy/gameoflife/GameRunner.java
@@ -15,7 +15,7 @@ public class GameRunner {
         this.scanner = scanner;
     }
 
-    public void startGame(int m, int n, int seedPercentage) {
+    public void startGame(int m, int n, double seedPercentage) {
         try {
             grid.seedRandomCells(m, n, seedPercentage);
             while (true) {
